WARNING: Although these changes don't show any errors, do not pull from this branch until further notice.
Testing out online code samples for how parsing from a database works.
A lot of this is to familiarize myself with the steps it would take to parse the data from the Chromesthesia database to the app. The tutorial I am working through displays information as a table on the android device.